### 2021-06-10 #3357 @whedon
[PRE REVIEW]: mpi4jax: Zero-copy MPI communication of JAX arrays
**Submitting author:** @dionhaefner (<a href="http://orcid.org/0000-0002-4465-7317">Dion Häfner</a>)
**Repository:** <a href="https://github.com/mpi4jax/mpi4jax" target ="_blank">https://github.com/mpi4jax/mpi4jax</a>
**Version:** v0.3.0
