### 2021-06-11 b0ff373
add JOSS paper from mpi4jax/mpi4jax-joss
