+++ paper.md
@@ -1,0 +1,218 @@
+---
+title: 'mpi4jax: Zero-copy MPI communication of JAX arrays'
+tags:
+  - Python
+  - JAX
+  - MPI
+  - high performance computing
+  - parallel computing
+
+authors:
+  - name: Dion Häfner
+    orcid: 0000-0002-4465-7317
+    affiliation: 1
+  - name: Filippo Vicentini^[Contributed equally, order determined by coin flip.]
+    affiliation: 2
+
+affiliations:
+ - name: Niels Bohr Institute, University of Copenhagen, Copenhagen, Denmark
+   index: 1
+ - name: Institute of Physics, École Polytechnique Fédérale de Lausanne (EPFL), CH-1015 Lausanne, Switzerland
+   index: 2
+
+date: 10 June 2021
+bibliography: paper.bib
+
+---
+
+# Summary
+
+The tensor framework JAX [@jax] combines expressivity and performance while retaining an accessible pure Python interface.
+Expressivity is achieved by treating functions as first-class objects, while efficiency is obtained by compiling to machine code just-ahead-of-time.
+
+However, machine learning and (high-performance) scientific computing are often conducted on different hardware stacks: Machine learning is typically done on few highly parallel units (GPUs or TPUs) connected to a single host CPU, while scientific models tend to run on clusters of dozens to thousands of CPUs.
+Unfortunately, support from JAX and the underlying compiler XLA is more mature in the former case.
+Notably, there is so far no built-in solution to communicate data between different nodes that is as sophisticated as the widely used MPI (Message Passing Interface) libraries [@mpistandard].
+
+We attempt to fill this gap and introduce `mpi4jax`, a Python library bringing first-class support for the most important MPI operations to JAX.
+We achieve this by defining a set of primitive functions matching MPI's operations, instructing JAX how to transform them and providing a native implementation to execute them.
+This has the result that users can communicate arbitrary JAX data without performance or usability penalties.
+In particular, `mpi4jax` is able to communicate data without copying from CPU and GPU memory (if built against a CUDA-aware MPI library) between one or multiple hosts (e.g. via an Infiniband network on a cluster).
+
+This also means that existing applications using e.g. NumPy and `mpi4py` can be ported seamlessly to the JAX ecosystem for potentially significant performance gains.
+
+# Statement of Need
+
+For decades, high-performance computing has been done primarily in low-level programming languages like Fortran or C.
+But the ubiquity of Python is starting to spill into this domain as well, thanks to its strong library ecosystem and wide adoption throughout the sciences.
+
+With a combination of NumPy [@numpy] and `mpi4py` [@mpi4py], Python users can already build massively parallel applications without delving into low-level programming languages, which is often advantageous when human time is more valuable than computing time. 
+However, such high-level frameworks are not always able to achieve peak performance, especially in more _niche_ workloads.
+
+Google's JAX library leverages the XLA compiler and supports just-in-time compilation (JIT) of (a subset of) Python code to XLA primitives. [The result is highly competitive performance on both CPU and GPU](https://github.com/dionhaefner/pyhpc-benchmarks) [@pyhpc-benchmarks]. 
+This approach, conceptually similar to Julia's deferred compilation model [@Bezanson2017], achieves low-level performance in a high-level language. With a strong performance baseline on single devices, the only thing missing is easy scalability to massively parallel hardware stacks, which we supply here.
+
+Two real-world use cases for `mpi4jax` are the ocean model Veros [@hafner_veros_2018] and the machine learning toolkit for many-body quantum systems NetKet [@carleo_netket_2019]:
+
[…]
