+++ paper.bib
@@ -1,0 +1,168 @@
+@article{hafner_veros_2018,
+  title    = {Veros v0.1 – a fast and versatile ocean simulator in pure {Python}},
+  volume   = {11},
+  issn     = {1991-959X},
+  url      = {https://gmd.copernicus.org/articles/11/3299/2018/},
+  doi      = {10.5194/gmd-11-3299-2018},
+  abstract = {{\textless}p{\textgreater}{\textless}strong class="journal-contentHeaderColor"{\textgreater}Abstract.{\textless}/strong{\textgreater} A general circulation ocean model is translated from Fortran to Python. Its code structure is optimized to exploit available Python utilities, remove simulation bottlenecks, and comply with modern best practices. Furthermore, support for Bohrium is added, a framework that provides a just-in-time compiler for array operations and that supports parallel execution on both CPU and GPU targets.{\textless}/p{\textgreater} {\textless}p{\textgreater}For applications containing more than a million grid elements, such as a typical {\textless}span class="inline-formula"{\textgreater}{\textless}math xmlns="http://www.w3.org/1998/Math/MathML" id="M1" display="inline" overflow="scroll" dspmath="mathml"{\textgreater}{\textless}mrow{\textgreater}{\textless}msup{\textgreater}{\textless}mn mathvariant="normal"{\textgreater}1{\textless}/mn{\textgreater}{\textless}mo{\textgreater}∘{\textless}/mo{\textgreater}{\textless}/msup{\textgreater}{\textless}mo{\textgreater}×{\textless}/mo{\textgreater}{\textless}msup{\textgreater}{\textless}mn mathvariant="normal"{\textgreater}1{\textless}/mn{\textgreater}{\textless}mo{\textgreater}∘{\textless}/mo{\textgreater}{\textless}/msup{\textgreater}{\textless}/mrow{\textgreater}{\textless}/math{\textgreater}{\textless}span{\textgreater}{\textless}svg:svg xmlns:svg="http://www.w3.org/2000/svg" width="34pt" height="11pt" class="svg-formula" dspmath="mathimg" md5hash="cb533c0820352f32a7ca953bf25ffe4a"{\textgreater}{\textless}svg:image xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="gmd-11-3299-2018-ie00001.svg" width="34pt" height="11pt" src="gmd-11-3299-2018-ie00001.png"/{\textgreater}{\textless}/svg:svg{\textgreater}{\textless}/span{\textgreater}{\textless}/span{\textgreater} horizontal resolution global ocean model, Veros is approximately half as fast as the MPI-parallelized Fortran base code on 24 CPUs and as fast as the Fortran reference when running on a high-end GPU. By replacing the original conjugate gradient stream function solver with a solver from the pyAMG Python package, this particular subroutine outperforms the corresponding Fortran version by up to 1 order of magnitude.{\textless}/p{\textgreater} {\textless}p{\textgreater}The study is concluded with a simple application in which the North Atlantic wave response to a Southern Ocean wind perturbation is investigated. It is found that even in a realistic setting the phase speeds of boundary waves matched the expectations based on theory and idealized models.{\textless}/p{\textgreater}},
+  language = {English},
+  number   = {8},
+  urldate  = {2021-03-19},
+  journal  = {Geoscientific Model Development},
+  author   = {Häfner, Dion and Jacobsen, René Løwe and Eden, Carsten and Kristensen, Mads R. B. and Jochum, Markus and Nuterman, Roman and Vinter, Brian},
+  month    = aug,
+  year     = {2018},
+  note     = {Publisher: Copernicus GmbH},
+  pages    = {3299--3312},
[…]
