+++ README.md
@@ -1,0 +1,3 @@
+Repository holding the upcoming JOSS paper for [`mpi4jax`](https://github.com/PhilipVinc/mpi4jax).
+
+Compile a preview here: https://whedon.theoj.org/
