### 2021-06-11 d8fc51f
fixes for jax 0.2.14
