+++ mpi4jax/_src/collective_ops/gather.py
@@ -198,7 +198,7 @@ def mpi_gather_abstract_eval(x, token, root, comm):
 
     return (
         abstract_arrays.ShapedArray(out_shape, x.dtype),
-        abstract_arrays.abstract_token,
+        core.abstract_token,
     )
 
 
