+++ mpi4jax/_src/collective_ops/allgather.py
@@ -169,7 +169,7 @@ def mpi_allgather_abstract_eval(x, token, comm):
     out_shape = (size, *x.shape)
     return (
         abstract_arrays.ShapedArray(out_shape, x.dtype),
-        abstract_arrays.abstract_token,
+        core.abstract_token,
     )
 
 
