+++ mpi4jax/_src/collective_ops/recv.py
@@ -179,7 +179,7 @@ def mpi_recv_xla_encode_gpu(c, x, token, source, tag, comm, status):
 def mpi_recv_abstract_eval(xs, token, source, tag, comm, status):
     return (
         abstract_arrays.ShapedArray(xs.shape, xs.dtype),
-        abstract_arrays.abstract_token,
+        core.abstract_token,
     )
 
 
