### 2021-06-11 b78c859
set minimum jax version
0.2.10 is the first version compatible with jaxlib 0.1.62
