### 2021-06-11 #3357 @danielskatz
The preview service seems to hang on it too, so my guess is there's something subtly wrong in the paper/references
