### 2021-06-24 #3357 @kellyrowland
Thanks! I'm going to reach out to a few folks further down the list to spread out reviewing opportunities, but the expertise list here is a big help.

Hi @csadorf and @1313e 👋 would you be available and willing to review this JOSS submission?
