+++ .github/workflows/mpi-tests.yml
@@ -77,3 +77,7 @@ jobs:
           else
             mpirun -host localhost:2 -v -np 2 pytest .
           fi
+
+      - name: Debug with tmate on failure
+        if: ${{ failure() }}
+        uses: mxschmitt/action-tmate@v3
