### 2021-07-12 #3419 @1313e
@dionhaefner I have a few questions regarding authorship and the paper itself.
I am posting them here such that it is easier to track for others and to allow for feedback.

I wanted to ask why only one of the two of you is stated as being an author of the software in the `setup.py` file?
Shouldn't that be both?

The order of authors on the paper is not the same as the order of the contributors (yes, I can see how made more contributions in terms of lines).
Given how important first author usually is, I want to make sure it is done correctly.
So, can you confirm that the order is correct?

There is a third person named `inailuig`, who despite only having 2 contributions, does seem to have made a proper impact on the package.
Is there a reason this person is not included as an author on the paper?

The paper itself seems to be a bit too long in my opinion.
Please correct me if I am wrong @kellyrowland, but the point of a JOSS paper is to basically give a summary or abstract to your software, just to make software easily citable.
The documentation should actually provide all the details and information required for using the package, whereas the package more gives a general description to get others interested in it.
In this particular case, there are quite a few code snippets with explanations in the paper that are more well-suited for the online documentation.
Furthermore, I also cannot find these code snippets with explanations in the online documentation, so it might be a good idea to move them there.
