### 2021-07-12 mpi4jax/mpi4jax#93 @1313e
[JOSS Review] Test requirements
This issue is part of the JOSS Review in openjournals/joss-reviews#3419.

I would like to execute the tests in `mpi4jax`, which use `pytest`, on my desktop, but there are no instructions nor requirements available for performing these tests.
Could a list of requirements be added, either as a file that can be installed with `pip install -r requirements_test.txt`, or simply adding it to the `setup()` function in the `setup.py` file?
