### 2021-07-12 mpi4jax/mpi4jax#94 @1313e
[JOSS Review] Community Guidelines
This issue is part of the JOSS Review in openjournals/joss-reviews#3419.

I am currently looking for a section in either the README or the online documentation that discusses how others can contribute to this code.
However, besides the very small section in the README mentioning that this package uses pre-commit hooks, there is nothing on this.
Can a section in either the README or the online documentation (preferably both) be added that describes how others can contribute to `mpi4jax`?
This can be a very small section simply stating that people should open issues when having problems, and use a PR to a specific branch for adding features.

PS: Having issue and/or PR templates also really helps with that.
