### 2021-07-12 mpi4jax/mpi4jax#93 @PhilipVinc
@1313e thanks for your review.

@dionhaefner is it ok if we go with adding an `extras_require={'dev':[pytest, black, precommit...]}` so that if one does `pip install mpi4jax[dev]` then everything will work out of the box?

PyTest is already set up to run the tests by simply invoking `pytest` so we might need to write in the documentation somewhere that you are supposed to invoke `pytest && mpirun -np 2 pytest`
