### 2021-07-12 mpi4jax/mpi4jax#93 @dionhaefner
Yes, sounds good. So if you take care of the extras, I can expand the README a bit with a mini developer & contribution guide.
