### 2021-07-12 mpi4jax/mpi4jax#93 @PhilipVinc
readme or a section in the documentation?
