### 2021-07-12 #3419 @dionhaefner
Thank you very much for your comments @1313e. Answers inline, and I will address your issues on the `mpi4jax` repo soon.

> I wanted to ask why only one of the two of you is stated as being an author of the software in the setup.py file?
Shouldn't that be both?
>
> The order of authors on the paper is not the same as the order of the contributors (yes, I can see how made more contributions in terms of lines).
Given how important first author usually is, I want to make sure it is done correctly.
So, can you confirm that the order is correct?

We flipped a coin for first authorship of the paper, as also indicated in the author list. We feel that both have contributed equally. So the order is correct, yes.

In case of the README and `setup.py`, we think that full attribution and first-authorship as not nearly as important as on an academic paper. The README just lists contributors in chronological order. But I actually think this list is redundant, as it's already shown in the GitHub UI in the sidebar. So I will remove the contributors section from the README.

In case of `setup.py`, this is often just the original author or main contact for questions. In my experience it is not meant to give an exhaustive list.

> There is a third person named inailuig, who despite only having 2 contributions, does seem to have made a proper impact on the package.
Is there a reason this person is not included as an author on the paper?

We value every contribution, but we feel that the bar for academic authorship should be a little bit higher than a one-time code contribution. If I interpret the JOSS guidelines correctly, authorship questions should be settled through "common sense", and we do feel that the author list is an accurate reflection of the work that went into the package.

> The paper itself seems to be a bit too long in my opinion.
Please correct me if I am wrong @kellyrowland, but the point of a JOSS paper is to basically give a summary or abstract to your software, just to make software easily citable.
The documentation should actually provide all the details and information required for using the package, whereas the package more gives a general description to get others interested in it.
In this particular case, there are quite a few code snippets with explanations in the paper that are more well-suited for the online documentation.
Furthermore, I also cannot find these code snippets with explanations in the online documentation, so it might be a good idea to move them there.

I take it you refer to the second half of the "Implementation" section? We use these code snippets as a device to explain why we need to go out of our way with token management, but if you think they are distracting we could remove them. There are similar examples [in the corresponding part of the docs](https://mpi4jax.readthedocs.io/en/latest/sharp-bits.html#token-management).

If you mean the shallow water application, that one is [also part of the docs](https://mpi4jax.readthedocs.io/en/latest/shallow-water.html). Same applies here, if you think the code is distracting we can remove it and refer to the docs, and just include the benchmark timings and the figure.

> Finally, the paper mentions in the Outlook that there are quite a few more functionalities that must be implemented for the package.
Isn't it maybe a bit too early to already publish it then, even if it is already useful to others (given the number of stars, I can see that it is being used)?

This seems to be a misunderstanding; there is nothing that *must* be implemented for the package to be useful. There are 2 major use cases for `mpi4jax`:

1. To use MPI primitives in JITed code without overhead on CPU and GPU.
2. To do distributed autograd, where you differentiate through communication calls.

Use case 1 is mature and fully supported, and this is what we describe in the paper (and, as you noted, is already being used in the wild). Use case 2 is not really supported, but might be in the future *(but it could also turn out to be technically impossible)*. I think it is customary to extrapolate a bit into the future in the last section, also to inspire potential contributors. If you prefer that we stick to what is already supported we can remove it.
