### 2021-07-12 mpi4jax/mpi4jax#93 @dionhaefner
Sure, I'll add it to the docs then.
