### 2021-07-12 d3986c2
update precommit versions and dev extras
