### 2021-07-12 24ea6fb
Update setup.py
Co-authored-by: Dion Häfner <dion.haefner@nbi.ku.dk>
