### 2021-07-12 #3419 @1313e
@dionhaefner

> We value every contribution, but we feel that the bar for academic authorship should be a little bit higher than a one-time code contribution. If I interpret the JOSS guidelines correctly, authorship questions should be settled through "common sense", and we do feel that the author list is an accurate reflection of the work that went into the package.

Okay, good.
I just wanted to make sure they weren't simply forgotten about.

> I take it you refer to the second half of the "Implementation" section? We use these code snippets as a device to explain why we need to go out of our way with token management, but if you think they are distracting we could remove them. There are similar examples [in the corresponding part of the docs](https://mpi4jax.readthedocs.io/en/latest/sharp-bits.html#token-management).

Yes, I am referring to that part.
I am not saying it isn't useful, because it most definitely is.
It is just a lot of information that I feel is a bit too extensive for the paper itself.
But again, if the editor is happy with it, then I am as well.

> This seems to be a misunderstanding; there is nothing that _must_ be implemented for the package to be useful. There are 2 major use cases for `mpi4jax`:
>
> 1. To use MPI primitives in JITed code without overhead on CPU and GPU.
> 2. To do distributed autograd, where you differentiate through communication calls.
>
> Use case 1 is mature and fully supported, and this is what we describe in the paper (and, as you noted, is already being used in the wild). Use case 2 is not really supported, but might be in the future _(but it could also turn out to be technically impossible)_. I think it is customary to extrapolate a bit into the future in the last section, also to inspire potential contributors. If you prefer that we stick to what is already supported we can remove it.

Please note that I am not saying that the package isn't useful or feature complete.
On the contrary actually.
I am more pointing out that more functionality seems to be planned for the future, as you mention as well, and to me it seemed to be quite a bit.
I was therefore simply wondering whether it would be a good idea to publish now instead of later, when the paper can also introduce those added functionalities.

Hope my comments didn't feel too direct.
I actually really like the package and made me interested in JAX and mpi4jax as well. :)
