### 2021-07-12 #3419 @dionhaefner
Thanks for the clarification! Sounds like we are on the same page. I think your comments were perfectly appropriate.

> I was therefore simply wondering whether it would be a good idea to publish now instead of later, when the paper can also introduce those added functionalities.

We have considered that, but given that we don't know whether autgrad-through-MPI will ever be part of `mpi4jax` (either because we don't have the resources or because of fundamental technical obstacles, perhaps both), we decided that it was a good time to publish now.
