### 2021-07-12 f66f193
Add [dev] to extras_require (#95)
* update precommit versions and dev extras

* Update setup.py

Co-authored-by: Dion Häfner <dion.haefner@nbi.ku.dk>

Co-authored-by: Dion Häfner <dion.haefner@nbi.ku.dk>
