### 2021-07-12 #3419 @PhilipVinc
> We flipped a coin for first authorship of the paper, as also indicated in the author list. We feel that both have contributed equally. So the order is correct, yes.

Bt the way, @dionhaefner , I noticed just now, but in my field it is customary to put the asterisk of "both authors contributed equally" on both authors (not only the second). So maybe we can address that?

> Finally, the paper mentions in the Outlook that there are quite a few more functionalities that must be implemented for the package.
Isn't it maybe a bit too early to already publish it then, even if it is already useful to others (given the number of stars, I can see that it is being used)?

I am writing an article for [netket](https://github.com/netket/netket) to be submitted in a few months and i would like to cite `mpi4jax` because NetKet could not work without it. I believe `mpi4jax` is now mature and battle tested by netket users enough to warrant a paper.
(In fact, `mpi4jax` was born as a subpackage in `netket` and was soon moved out once i started discussing with @dionhaefner because it looked like it could gain more traction.)

About not being complete: me and dion see different value in `mpi4jax`.
He sees part 1, the collective communication, as the most important part, while i am mostly interested in part 2, that is autodiff through mpi calls, which is what I need in NetKet.
As he said, part 1 is feature-complete (bar a few exotic mpi calls, which i don't think warrant stalling the article) while part 2 is not. Many things do work, even if we don't mention them in the article (mainly collective communication), but some (send/recv) do  with very peculiar limitations (like only in reverse mode, or only in forward mode) due to limitations in jax that even I am not 100% sure if they can be overcome or not.

For that reason we agreed to focus the JOSS paper on part 1 (MPI), and mention part 2 (AD) as something that can be fully fleshed out in the future, even if we do already support AD in the more common use-cases.
