### 2021-07-12 a677598
add footnote to both authors
