### 2021-07-12 #3419 @dionhaefner
> Bt the way, @dionhaefner , I noticed just now, but in my field it is customary to put the asterisk of "both authors contributed equally" on both authors (not only the second). So maybe we can address that?

We had that in the first draft, but at some point it disappeared, so I assumed you removed it :) Anyhow, it's back.
