### 2021-07-12 mpi4jax/mpi4jax#94 @PhilipVinc
Regarding the Issue template, @dionhaefner , in NetKet we recently implemented a small utility to help users debug and provide us decent informations (since we were seeing lots of MPI-related bug reports, mainly due to misconfiguration on SLURM clusters).
The plugin is something liek

```python
netket on   (git)-[pv/expect|bisect]- [$] via netket_env via 🐍 system took 2s
➜ python -m netket.tools.info
====================================================
==         NetKet Diagnostic Informations         ==
====================================================
NetKet version       : 3.0b3.dev52+gc3797f66.d20210629

# Python
  - implementation   : CPython
  - version          : 3.8.10
  - distribution     : GCC 11.1.1 20210428 (Red Hat 11.1.1-1)
  - path             : /home/filippovicentini/Documents/pythonenvs/netket_env/bin/python

# Host informations
  - System           : Linux-5.12.13-300.fc34.x86_64-x86_64-with-glibc2.2.5
  - Architecture     : x86_64
  - AVX              : True
  - AVX2             : True
  - Cores            : 32

# NetKet dependencies
  - numpy            : 1.21.0
  - jaxlib           : 0.1.68
  - jax              : 0.2.16
  - flax             : 0.3.4
  - optax            : 0.0.6
  - numba            : 0.53.1
  - mpi4py           : 3.0.3
  - mpi4jax          : 0.3.1
  - netket           : 3.0b3.dev52+gc3797f66.d20210629

# Jax
WARNING:absl:No GPU/TPU found, falling back to CPU. (Set TF_CPP_MIN_LOG_LEVEL=0 and rerun for more info.)
  - backends         : ['cpu']
    - cpu            : [<jaxlib.xla_extension.Device object at 0x7f4c5ce03db0>]

# MPI4JAX
  - HAS_GPU_EXT      : True

# MPI
  - mpi4py
  - MPICC            : /usr/lib64/mpich/bin/mpicc
  - MPI link flags   : ['-L/usr/lib64/mpich/lib']
    - MPI version    : (3, 1)
    - MPI library_version : MPICH Version:	3.4.1
MPICH Release date:	Fri Jan 22 14:17:48 CST 2021
MPICH ABI:	13:10:1
MPICH Device:	ch3:nemesis
MPICH configure:	--build=x86_64-redhat-linux-gnu --host=x86_64-redhat-linux-gnu --program-prefix= --disable-dependency-tracking --prefix=/usr --exec-prefix=/usr --bindir=/usr/bin --sbindir=/usr/sbin --sysconfdir=/etc --datadir=/usr/share --includedir=/usr/include --libdir=/usr/lib64 --libexecdir=/usr/libexec --localstatedir=/var --sharedstatedir=/var/lib --mandir=/usr/share/man --infodir=/usr/share/info --enable-sharedlibs=gcc --enable-shared --enable-static=no --enable-lib-depend --disable-rpath --disable-silent-rules --enable-fortran --with-gnu-ld --with-device=ch3:nemesis --with-pm=hydra:gforker --includedir=/usr/include/mpich-x86_64 --bindir=/usr/lib64/mpich/bin --libdir=/usr/lib64/mpich/lib --datadir=/usr/share/mpich --mandir=/usr/share/man/mpich-x86_64 --docdir=/usr/share/mpich/doc --htmldir=/usr/share/mpich/doc --with-hwloc-prefix=system FFLAGS=-O2 -flto=auto -ffat-lto-objects -fexceptions -g -grecord-gcc-switches -pipe -Wall -Werror=format-security -Wp,-D_FORTIFY_SOURCE=2 -Wp,-D_GLIBCXX_ASSERTIONS -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -fstack-protector-strong -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1 -m64 -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection -I/usr/lib64/gfortran/modules -fallow-argument-mismatch
MPICH CC:	gcc -O2 -flto=auto -ffat-lto-objects -fexceptions -g -grecord-gcc-switches -pipe -Wall -Werror=format-security -Wp,-D_FORTIFY_SOURCE=2 -Wp,-D_GLIBCXX_ASSERTIONS -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -fstack-protector-strong -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1  -m64  -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection   -O2
MPICH CXX:	g++ -O2 -flto=auto -ffat-lto-objects -fexceptions -g -grecord-gcc-switches -pipe -Wall -Werror=format-security -Wp,-D_FORTIFY_SOURCE=2 -Wp,-D_GLIBCXX_ASSERTIONS -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -fstack-protector-strong -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1  -m64  -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection  -O2
MPICH F77:	gfortran -O2 -flto=auto -ffat-lto-objects -fexceptions -g -grecord-gcc-switches -pipe -Wall -Werror=format-security -Wp,-D_FORTIFY_SOURCE=2 -Wp,-D_GLIBCXX_ASSERTIONS -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -fstack-protector-strong -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1  -m64  -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection -I/usr/lib64/gfortran/modules -fallow-argument-mismatch  -O2
MPICH FC:	gfortran -O2 -flto=auto -ffat-lto-objects -fexceptions -g -grecord-gcc-switches -pipe -Wall -Werror=format-security -Wp,-D_FORTIFY_SOURCE=2 -Wp,-D_GLIBCXX_ASSERTIONS -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -fstack-protector-strong -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1  -m64  -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection -I/usr/lib64/gfortran/modules  -O2

which: no mpicc in (/home/filippovicentini/Documents/pythonenvs/netket_env/bin:/home/filippovicentini/Documents/pythonenvs/netket_env/bin:/home/filippovicentini/.zplug/bin:/usr/local/sbin:/home/filippovicentini/.julia/bin:/home/filippovicentini/bin/:/usr/local/cuda/bin:/home/filippovicentini/.pyenv/bin:/home/filippovicentini/.local/bin:/home/filippovicentini/bin:/usr/local/bin:/usr/bin:/usr/sbin)
which: no mpirun in (/home/filippovicentini/Documents/pythonenvs/netket_env/bin:/home/filippovicentini/Documents/pythonenvs/netket_env/bin:/home/filippovicentini/.zplug/bin:/usr/local/sbin:/home/filippovicentini/.julia/bin:/home/filippovicentini/bin/:/usr/local/cuda/bin:/home/filippovicentini/.pyenv/bin:/home/filippovicentini/.local/bin:/home/filippovicentini/bin:/usr/local/bin:/usr/bin:/usr/sbin)
which: no mpiexec in (/home/filippovicentini/Documents/pythonenvs/netket_env/bin:/home/filippovicentini/Documents/pythonenvs/netket_env/bin:/home/filippovicentini/.zplug/bin:/usr/local/sbin:/home/filippovicentini/.julia/bin:/home/filippovicentini/bin/:/usr/local/cuda/bin:/home/filippovicentini/.pyenv/bin:/home/filippovicentini/.local/bin:/home/filippovicentini/bin:/usr/local/bin:/usr/bin:/usr/sbin)
  - global
    - MPICC          :
    - MPI link flags :

```

We could copy-paste the relevant parts from it and use it together with an issue template?
