### 2021-07-12 mpi4jax/mpi4jax#94 @dionhaefner
If you want to do this, sure. Otherwise I think it's also OK to just add a small section to the docs for now, and add an issue template when we need it (i.e., when we regularly get bug reports).
