### 2021-07-12 fe4f4d9
add developer guide; remove contributors from readme
