### 2021-07-12 64f6408
add developer guide; remove contributors from readme (#97)
