### 2021-07-12 mpi4jax/mpi4jax#93 @dionhaefner
How's this?
https://github.com/mpi4jax/mpi4jax#community-guidelines
https://mpi4jax.readthedocs.io/en/latest/developers.html
