### 2021-07-14 mpi4jax/mpi4jax#93 @1313e
However, it seems that `pip install -e .[dev]` does not install anything extra, as `pip` does not understand that command:
```WARNING: mpi4jax 0.3.1+2.g64f6408 does not provide the extra 'dev'```.
