### 2021-07-14 c26cfd5
fix extras_require (#93)
