### 2021-07-14 mpi4jax/mpi4jax#93 @1313e
Might be a good idea to install the package using the described way from the online documentation in GitHub Actions, to make sure that there is a minimal difference between the user way and automated way of testing.
