### 2021-07-14 2e85496
do not require tqdm in example; use dev extra in CI
