+++ .github/workflows/mpi-tests.yml
@@ -53,10 +53,8 @@ jobs:
           fi
 
           python -m pip install --upgrade pip
-          pip install wheel
-          pip install pytest numpy jax jaxlib mpi4py tqdm
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-          pip install .
+          pip install jax jaxlib
+          pip install .[dev]
 
       - name: Single-execution test with pytest
         run: |
