### 2021-07-14 81a3f29
CI cleanup (#98)
