+++ pyproject.toml
@@ -1,3 +1,9 @@
+[tool.pytest.ini_options]
+testpaths = [
+    "tests",
+]
+
+
 [tool.coverage.run]
 branch = true
 parallel = true
