+++ .github/workflows/mpi-tests.yml
@@ -53,24 +53,18 @@ jobs:
           fi
 
           python -m pip install --upgrade pip
-          pip install wheel
-          pip install pytest numpy jax jaxlib mpi4py tqdm
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-          pip install .
+          pip install jax jaxlib
+          pip install .[dev]
 
       - name: Single-execution test with pytest
         run: |
           if [ ${{ matrix.mpi }} == 'intelmpi' ]; then
             source ${HOME}/intel/compilers_and_libraries/linux/mpi/intel64/bin/mpivars.sh -ofi_internal=1 release
           fi
-
-          cd tests
           pytest .
 
       - name: mpirun test with pytest
         run: |
-          cd tests
-
           if [ ${{ matrix.mpi }} == 'intelmpi' ]; then
             source ${HOME}/intel/compilers_and_libraries/linux/mpi/intel64/bin/mpivars.sh -ofi_internal=1 release
             mpirun -v -np 2 pytest .
