+++ .github/workflows/python-publish.yml
@@ -35,7 +35,6 @@ jobs:
         env:
           TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
           TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-
         run: |
           python setup.py sdist
           twine upload dist/*
