+++ .github/workflows/covecov-coverage.yml
@@ -40,10 +40,8 @@ jobs:
       - name: Pip install python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install wheel
-          pip install pytest numpy jax jaxlib mpi4py tqdm cython coverage
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-          pip install -v -e .
+          pip install jax jaxlib
+          pip install -e .[dev]
 
       - name: Single-execution test with pytest and coverage
         run: |
