### 2021-07-16 #3419 @csadorf
@kellyrowland I did not realize that one of the authors is affiliated with institution I am employed by: École Polytechnique Fédérale de Lausanne (EPFL). According to the COI Policy, this constitutes a COI. I have not personally interacted or collaborated with Filippo Vicentini or their host laboratory. I am going to continue with the review process, but will wait on confirmation on whether the COI is waived or not before submitting any comments.
