### 2021-07-20 mpi4jax/mpi4jax#100 @1313e
[JOSS Review] Mention `mpi4py` being required
This issue is part of the JOSS Review in openjournals/joss-reviews#3419.

In order for `mpi4jax` to work, `mpi4py` is required, which is obvious.
However, for some, it might be logical to assume that `mpi4jax` will install `mpi4py` automatically correctly.
The problem however is that installing `mpi4py` is usually a bit of a pain to do correctly and rarely works properly, so it is much better to do it manually.
Therefore, I would advise putting in the README at the basic installation instructions that `mpi4py` must be installed already.

It might also be a good idea to make this clearer in the online documentation installation instructions.
