### 2021-07-20 mpi4jax/mpi4jax#101 @1313e
[JOSS Review] Add `jaxlib` to requirements
This issue is part of the JOSS Review in openjournals/joss-reviews#3419.

`mpi4jax` requires `jax` in order to function, which is very logical.
However, `jax` requires `jaxlib` to function properly.
Given that, for some reason, `jax` does not automatically install `jaxlib`, either `jax` or `mpi4jax` must add this requirement to the requirements list.
Otherwise `mpi4jax` cannot be installed properly using a single install command.
