### 2021-07-20 mpi4jax/mpi4jax#102 @1313e
[JOSS Review] Tests cannot be executed properly
This issue is part of the JOSS Review in openjournals/joss-reviews#3419.

I have installed `mpi4jax` properly (I think), but when I want to run the tests, the process crashes at `tests/test_examples.py::test_shallow_water_demo` with no errors whatsoever.
Normally I would say that this is an indication of a SegmentationFault in C, but `mpi4jax` does not use C.

Below is what my actual output is
```
x1313e@CyborgU:~/Bureaublad/mpi4jax$ pytest -vvv
=============================================================================================== test session starts ===============================================================================================
platform linux -- Python 3.7.4, pytest-6.2.4, py-1.10.0, pluggy-0.13.1 -- /home/x1313e/.conda/envs/mpi4jax/bin/python
cachedir: .pytest_cache
MPI vendor: ('MPICH', (3, 3, 2))
MPI rank: 0
MPI size: 1
rootdir: /home/x1313e/Bureaublad/mpi4jax, configfile: pyproject.toml, testpaths: tests
plugins: cov-2.12.1
collected 108 items

tests/test_decorators.py::test_envvar_parsing PASSED                                                                                                                                                        [  0%]
tests/test_decorators.py::test_missing_omnistaging PASSED                                                                                                                                                   [  1%]
tests/test_decorators.py::test_ensure_gpu_ext PASSED                                                                                                                                                        [  2%]
tests/test_examples.py::test_shallow_water_demo
```
The fourth test never runs, as I get control back of the terminal.

If I disable that test to see if the others work, I get the same issue with the test called `tests/collective_ops/test_allgather.py::test_allgather`, so it seems there are multiple of these.

Any ideas what is going wrong here?
