### 2021-07-20 mpi4jax/mpi4jax#101 @dionhaefner
This is by design. There are different versions of jaxlib for CPU and GPU, and I guess the JAX people want users to make a conscious choice of which one to use. We don't want to interfere with that and found it best to stick to what JAX is doing. Given that `mpi4jax` users are probably *already* active JAX users, we don't think this is a huge problem in practice. Let us know if you thought that isn't documented properly (I think we mention installing jaxlib in the docs).

BTW, according to the official JAX instructions they now *do* include jaxlib with `pip install jax`: https://github.com/google/jax/#installation
