### 2021-07-20 mpi4jax/mpi4jax#102 @dionhaefner
It looks like a broken MPI or `mpi4py` installation. @PhilipVinc has more experience with debugging broken installs, so maybe he can weigh in :)

I'm on vacation for 3 weeks now, if this is still standing I can take care of it.

Anyhow, more information on how you installed MPI, mpi4py, and mpi4jax is probably in order. Could also be a duplicate of https://github.com/mpi4jax/mpi4jax/issues/91, so please make sure you have `jaxlib>=0.1.69`.
