### 2021-07-20 mpi4jax/mpi4jax#102 @dionhaefner
It could also be a duplicate of #86 if you installed mpi4py before mpi4jax.
