### 2021-07-20 mpi4jax/mpi4jax#102 @PhilipVinc
That bug should only manifests if you have non multiple of 4 hardware threads, IIRC.
To rule it out you should update to the last jaxlib release.
