### 2021-07-20 mpi4jax/mpi4jax#100 @dionhaefner
It should work fine most of the time, but yes there are pathologic cases where it doesn't work (e.g. globally pre-installed mpi4py). For this we have this section in the docs:

https://mpi4jax.readthedocs.io/en/latest/installation.html#selecting-the-mpi-distribution

I think the MPI shipped by mpi4py is OpenMPI-based, so this might also break when using a different vendor. Is this what you ran into @1313e ?

If @PhilipVinc doesn't beat me to it I can update the docs (to mention mpi4py earlier) when I'm back in August.
