### 2021-07-21 mpi4jax/mpi4jax#100 @1313e
The MPI shipped with `mpi4py` is usually MPICH2, not OpenMPI.
Anyway, simply mentioning that it is a good idea to install `mpi4py` manually before attempting to install `mpi4jax` in the README should be enough.
