### 2021-07-21 mpi4jax/mpi4jax#101 @1313e
> BTW, according to the official JAX instructions they now _do_ include jaxlib with `pip install jax`: https://github.com/google/jax/#installation

I installed `jax` yesterday and that did not install `jaxlib` automatically.

Anyway, simply mentioning that `jax` and `jaxlib` must be installed manually before installing `mpi4jax` in the README would be fine.
I am saying the README here, as many users only look at the README and not at the docs for installation instructions.

PS: `mpi4jax` does install `jax` automatically, as it has it as a requirement, but not `jaxlib`.
If I am going to be honest, it might actually be better to remove `jax` as the requirement, to force users to install `jax` (and thus `jaxlib`) themselves properly.
Because `jax` is currently automatically taken care of, it feels weird that `jaxlib` is not, regardless of what reasons Google has for not including it.
