### 2021-07-21 mpi4jax/mpi4jax#102 @1313e
I installed the entire package plus `jaxlib` yesterday from `pip` in a new environment, so I am assuming that those are the latest versions.

All packages were installed using `pip`, where I installed `mpi4py` before installing `mpi4jax`.
I can use `mpi4py` just fine, all the communication methods that I tried to use execute as I expect them to.
