### 2021-07-22 #3419 @kellyrowland
Hi @csadorf - thanks for bringing this up. I will try to find another reviewer in the interest of conservatism with respect to the COI policy. Having said that, if we're unable to find another reviewer, it does sound like there is enough separation in this case that we could waive it.
