### 2021-07-22 #3419 @kellyrowland
Hello @adam-m-jcbs @gonsie 👋 would either of you be available and willing to review this paper for JOSS?
