### 2021-07-28 #3419 @Himscipy
👋🏼  @kellyrowland,
I am available to review this JOSS submission. Thank you for reaching out.
