### 2021-08-05 mpi4jax/mpi4jax#102 @dionhaefner
I can't reproduce this. Which Linux flavor is this? How did you install MPICH?
