### 2021-08-05 mpi4jax/mpi4jax#102 @1313e
Ubuntu, and MPICH was installed automatically by `mpi4py`.
I know the latter is installed properly, as my own MPI scripts can be executed just fine
