### 2021-08-06 mpi4jax/mpi4jax#102 @dionhaefner
> MPICH was installed automatically by mpi4py.

This sounds to me like you are mixing conda and pip, which we specifically advise against. mpi4py only ships with MPICH on conda (or on Windows).

That being said, I'm still unable to reproduce. I'm trying with this Dockerfile:

```dockerfile
FROM continuumio/miniconda3:latest

RUN conda create -n mpi4jax "python<3.8" mpi4py gxx_linux-64

SHELL ["conda", "run", "-n", "mpi4jax", "/bin/bash", "-c"]

WORKDIR /opt
RUN git clone https://github.com/mpi4jax/mpi4jax.git

WORKDIR /opt/mpi4jax
RUN pip install jax jaxlib
RUN pip install -e .[dev]

ENTRYPOINT ["conda", "run", "-n", "mpi4jax", "/bin/bash", "-c"]
```

Build and run via

```bash
$ docker build -t mpi4jax .
$ docker run mpi4jax pytest
```

Output:

```
MPI vendor: ('MPICH', (3, 3, 2))
MPI rank: 0
MPI size: 1
rootdir: /opt/mpi4jax, configfile: pyproject.toml, testpaths: tests
plugins: cov-2.12.1
collected 108 items

tests/test_decorators.py ...                                             [  2%]
tests/test_examples.py .                                                 [  3%]
tests/test_flush.py .                                                    [  4%]
tests/test_invalid_jaxlib.py ............                                [ 15%]
tests/test_validation.py ....                                            [ 19%]
tests/collective_ops/test_allgather.py ....                              [ 23%]
tests/collective_ops/test_allreduce.py ...............                   [ 37%]
tests/collective_ops/test_allreduce_matvec.py ........................   [ 59%]
tests/collective_ops/test_alltoall.py ...                                [ 62%]
tests/collective_ops/test_barrier.py .                                   [ 62%]
tests/collective_ops/test_bcast.py ....                                  [ 66%]
tests/collective_ops/test_common.py ...                                  [ 69%]
tests/collective_ops/test_gather.py ....                                 [ 73%]
tests/collective_ops/test_reduce.py ....                                 [ 76%]
tests/collective_ops/test_scan.py ....                                   [ 80%]
tests/collective_ops/test_scatter.py ...                                 [ 83%]
tests/collective_ops/test_send_and_recv.py sssssss                       [ 89%]
tests/collective_ops/test_sendrecv.py sssssssssss                        [100%]

=============================== warnings summary ===============================
../conda/envs/mpi4jax/lib/python3.7/site-packages/flatbuffers/compat.py:19
  /opt/conda/envs/mpi4jax/lib/python3.7/site-packages/flatbuffers/compat.py:19: DeprecationWarning: the imp module is deprecated in favour of importlib; see the module's documentation for alternative uses
    import imp

tests/test_examples.py::test_shallow_water_demo
  /opt/mpi4jax/tests/../examples/shallow_water.py:35: UserWarning: Could not import tqdm, can't show progress bar
    warnings.warn("Could not import tqdm, can't show progress bar")

-- Docs: https://docs.pytest.org/en/stable/warnings.html
================= 90 passed, 18 skipped, 2 warnings in 14.19s ==================
```

Do you see anything you did differently from what is done in the Dockerfile?
