### 2021-08-06 mpi4jax/mpi4jax#102 @1313e
`mpi4py` installs MPICH automatically on both `conda` and `pip` if you have no valid MPI distribution available.
On Windows it does not do that, as only Microsoft MPI will work on Windows (I know, as I use both Windows and Linux).
I however tried using solely `pip` to install everything, and that gives me the exact same problem as when I install everything using `conda`.
The above in the Dockerfile is exactly how I installed `mpi4jax`.
