### 2021-08-06 mpi4jax/mpi4jax#102 @dionhaefner
> mpi4py installs MPICH automatically on both conda and pip if you have no valid MPI distribution available.

That doesn't seem right. Just [look at the PyPI page](https://pypi.org/project/mpi4py/#files), you can see that there's no wheels for Linux, only Windows. Installing via pip definitely doesn't give you `mpirun` or `mpicc` executables. I have verified this in a clean Docker image.

As a last attempt to debug this, could you try

```bash
$ pip uninstall mpi4jax
$ pip install -e .[dev] --no-build-isolation --no-cache
$ pytest
```

and see if it changes anything?
