### 2021-08-09 mpi4jax/mpi4jax#101 @dionhaefner
I agree that the current state of installing jaxlib manually is awkward, but I really don't see a better solution. IMO the error when running without jaxlib is obvious enough that this shouldn't be a major obstacles for new users (which we assume to mostly be JAX users anyway).

I've added a section to the README, feel free to reopen if you feel that this is still an issue and we'll work on a better solution.
