### 2021-08-09 mpi4jax/mpi4jax#100 @dionhaefner
I'm not sure whether I agree that it is generally good to install mpi4py beforehand. Anyway, I have added instructions for this case to README and docs. Turns out the installation works a bit differently when you choose to pre-install mpi4py.

Feel free to reopen if you feel that this isn't resolved.
