### 2021-08-09 mpi4jax/mpi4jax#101 @1313e
It is good now. :)
It is simply to take out any confusion with beginners.
Should always try to make it as easy as possible for them to get started, and leaving very little up for interpretation.
