### 2021-08-11 mpi4jax/mpi4jax#102 @1313e
Sorry for taking so long, had a lot of business to take care off on my end.

On this PC, yes, I have a valid MPI distribution, as I need it for my work.
However, as far as I know, Linux does not come with its own MPI distribution and I have never installed one myself.
I always get it through `mpi4py`.

I tried to execute the script you gave me above, but that results in the same immediate crash with no errors thrown at all.

Afterward, I tried doing it again, but this time in Docker and on a supercomputer.
It seems to work just fine on both, so either there is something funny going on on my PC or MPICH2 is being weird.
On the supercomputer, I installed `mpi4jax` the exact same way as I did on my personal PC (which also means throwing `pip` and `conda` together), and that worked fine.
So, I am happy to close this issue now, but this might be something to keep in mind if others encounter the same problem, as I have absolutely no idea what is different between the installation I used on my machine and the supercomputer.
