### 2021-08-11 mpi4jax/mpi4jax#102 @dionhaefner
Sure, let's leave it at that for now and see if it comes up again in the future.

> However, as far as I know, Linux does not come with its own MPI distribution and I have never installed one myself.

I usually just install OpenMPI via

```bash
$ sudo apt install libopenmpi-dev
```
