### 2021-08-16 #3419 @Himscipy
Hi  @dionhaefner,  Following are some comments regarding the submission :
1. The software tools wraps mpi4py library with jax, is a very interesting idea and tool such as this will be really beneficial for scientific community. With that said, I find that the paper doesn't describe findings the finding from the second example listed in the introduction section of the paper(NetKet).
2.  The paper does describe the scalability for the a 2D shallow water on ~6M grid, but doesn't clearly mention about the workload handled by each worker, and the mesh partition strategy a user can directly leverage from the tool. For the general application perspective, it's important to clearly describe this in the paper.
3. In addition if you also provide a 3D example case and the working mesh partitioning strategies, the usability of the tool can increase multifold. I am not aware of existing JAX based Numerical scientific solvers, if there are such existing codes, a impactful use case would be to show how your software tool can integrate with it seamlessly. The paper does mention about a  Jax based code "Veros" but did not reports the scalability and performance. In addition the "Veros " documentation does list mpi4Jax for distributed computing.
4. As the work talks about communication and distributed computing, for HPC community it is important to have code profiling and timing results from the profiler to demonstrate the applicability of the tool.
5. While the reported benchmarks are on small machine, I recommend to test the wrapper performance on large clusters as well and report their results and findings in the paper.
6. The "zero copy " mentioned in the paper title is never explained or elaborated in the paper. My recommendation would be to provide explanation about it in the paper.
