### 2021-08-16 #3419 @dionhaefner
Thanks for the review @Himscipy.

> 1. The software tools wraps mpi4py library with jax, is a very interesting idea and tool such as this will be really beneficial for scientific community. With that said, I find that the paper doesn't describe findings the finding from the second example listed in the introduction section of the paper(NetKet).

I'm not sure I follow. What kind of extra information would you expect? I'm sure @PhilipVinc would be happy to provide.

> 2. The paper does describe the scalability for the a 2D shallow water on ~6M grid, but doesn't clearly mention about the workload handled by each worker, and the mesh partition strategy a user can directly leverage from the tool. For the general application perspective, it's important to clearly describe this in the paper.

I have no problem with mentioning the mesh partitioning and grid points per worker. I'll add it to the manuscript. I'm not sure though what you mean by "a user can directly leverage from the tool". All interaction with communicators is handled on the user side via mpi4py, so there's nothing we add or take away (see also next point).

> 3. In addition if you also provide a 3D example case and the working mesh partitioning strategies, the usability of the tool can increase multifold. I am not aware of existing JAX based Numerical scientific solvers, if there are such existing codes, a impactful use case would be to show how your software tool can integrate with it seamlessly. The paper does mention about a Jax based code "Veros" but did not reports the scalability and performance. In addition the "Veros " documentation does list mpi4Jax for distributed computing.

I agree that a 3D example would be great, but it's not exactly something I have "lying around", and since JAX is relatively new there's not a huge ecosystem to fall back to. If you have a simple 3D toy example I would be happy to see if we can make it work - that would definitely make a good addition to the docs.

I am the maintainer of Veros, and yes we do use `mpi4jax` to scale to small clusters and dozens of GPUs. I could add [a link to the Veros benchmarks page](https://veros.readthedocs.io/en/latest/more/benchmarks.html#varying-number-of-mpi-processes) to the paper?

> 4. As the work talks about communication and distributed computing, for HPC community it is important to have code profiling and timing results from the profiler to demonstrate the applicability of the tool.

We could try and quantify the overhead incurred by using mpi4jax by benchmarking it against mpi4py. Would that sound like a useful addition? Otherwise you'll need to be more specific *what* we should profile, since all actual communication happens inside mpi4py / MPI.

> 5. While the reported benchmarks are on small machine, I recommend to test the wrapper performance on large clusters as well and report their results and findings in the paper.

We benchmark Veros for up to 256 cores (see link above). Would it be sufficient to link to that, given that we also measure overhead vs mpi4py (4)? (Besides that, I don't have access to a *large* cluster.)

> 6. The "zero copy " mentioned in the paper title is never explained or elaborated in the paper. My recommendation would be to provide explanation about it in the paper.

Sure, I'll add a sentence.
