### 2021-08-17 #3419 @PhilipVinc
If I can add something...

>4. As the work talks about communication and distributed computing, for HPC community it is important to have code profiling and timing results from the profiler to demonstrate the applicability of the tool.

While I do agree that benchmarks in general are a usefull tool, i don't think they are really relevant here.
`mpi4jax` is literally just a wrapper+autodiff rules, as you recognise yourself in your question (5).
It just turns out that writing a wrapper for jax is not super-simple and we would like to be able to cite our work for it.

Properly benchmarking `mpi4jax` will, in the end, tell us what is the cost of a function call in XLA and how performant is the particular MPI implementation of the (super)computer used in the benchmark. Moreover, in any reasonably-sized workload this cost will be negligible compared to the MPI operation itself.

I don't think that this information is particularly useful.
(Addendum: It is also a moving target: customcall runtime cost has been halved in recent versions of jaxlib/XLA and depends on the library version the user has installed).
