### 2021-08-18 #3419 @dionhaefner
FWIW, I did some quick benchmarks and found that `mpi4jax` incurs an overhead of about 800ns (so way beyond noise level in any application I can think of).

I think that could be a nice result to add to the paper, to show that we actually achieve zero-copy communications with minimal overhead.
