### 2021-08-25 #3419 @Himscipy
`I'm not sure I follow. What kind of extra information would you expect? I'm sure @PhilipVinc would be happy to provide.`
I don't expect extra information, here I mean that in the introduction section you elucidate the NetKet example, but in the result section you only show the results of the shallow water equation. Providing the details of all the examples you list in the introduction section, will provide completeness to the paper. In addition, I didn't find any details on the NetKet in the documentation page also.

`I have no problem with mentioning the mesh partitioning and grid points per worker. I'll add it to the manuscript. I'm not sure though what you mean by "a user can directly leverage from the tool". All interaction with communicators is handled on the user side via mpi4py, so there's nothing we add or take away.`
- When I say user can directly leverage from the tool. I mean if there are mesh partitioning code/dependency provided directly  by Mpi4Jax, that can be directly use to divide the mesh among the worker. Hence I say details of what library/code base for mesh partitioning was used is important to mention to increase the usability of the tool.

`I think that could be a nice result to add to the paper, to show that we actually achieve zero-copy communications with minimal overhead.`
Please add that in the paper, this will further strengthen the tool usability.
