### 2021-08-30 fb16a75
Better logging (including timing) (#105)
