+++ mpi4jax/_src/xla_bridge/mpi_xla_bridge.pyx
@@ -1,4 +1,7 @@
 import sys
+import string
+import random
+from time import perf_counter
 
 from libc.stdint cimport uintptr_t
 
@@ -28,6 +31,7 @@ MPI_STATUS_IGNORE_ADDR = int(<uintptr_t>MPI_STATUS_IGNORE)
 # Logging
 #
 
+
 cdef bint PRINT_DEBUG = False
 
 
@@ -40,11 +44,20 @@ cpdef bint get_logging():
     return PRINT_DEBUG
 
 
-cdef inline void print_debug(unicode message, MPI_Comm comm) nogil:
+cdef inline void print_debug(unicode message, unicode rid, MPI_Comm comm) nogil:
     cdef int rank
     MPI_Comm_rank(comm, &rank)
     with gil:
-        print(f"r{rank} | {message}", flush=True)
+        print(f"r{rank} | {rid} | {message}", flush=True)
+
+
+cdef unicode random_id():
+    cdef unicode alphabet = (
+        string.ascii_lowercase
+        + string.ascii_uppercase
+        + string.digits
+    )
+    return ''.join(random.choices(alphabet, k=8))
 
 
 #
@@ -86,13 +99,18 @@ cdef void mpi_barrier(MPI_Comm comm) nogil:
 
     if PRINT_DEBUG:
         with gil:
-            print_debug(
-                "MPI_Barrier",
-                comm
-            )
+            rid = random_id()
+            print_debug(f"MPI_Barrier", rid, comm)
+            start = perf_counter()
 
     # MPI Call
     ierr = libmpi.MPI_Barrier(comm)
+
+    if PRINT_DEBUG:
+        with gil:
+            end = perf_counter()
+            print_debug(f"MPI_Barrier done with code {ierr} ({end - start:.2e}s)", rid, comm)
+
     abort_on_error(ierr, comm, u"Barrier")
 
 
@@ -103,28 +121,49 @@ cdef void mpi_allgather(void* sendbuf, int sendcount, MPI_Datatype sendtype,
 
     if PRINT_DEBUG:
         with gil:
+            rid = random_id()
             print_debug(
                 f"MPI_Allgather sending {sendcount}, receiving {recvcount} items",
+                rid,
                 comm
             )
+            start = perf_counter()
 
     # MPI Call
     ierr = libmpi.MPI_Allgather(
         sendbuf, sendcount, sendtype, recvbuf, recvcount, recvtype, comm
     )
+
+    if PRINT_DEBUG:
+        with gil:
+            end = perf_counter()
+            print_debug(
+                f"MPI_Allgather done with code {ierr} ({end - start:.2e}s)",
+                rid,
+                comm
+            )
+
     abort_on_error(ierr, comm, u"Allgather")
 
 
-cdef void mpi_allreduce(void* sendbuf, void* recvbuf, int nitems,
+cdef void mpi_allreduce(void* sendbuf, void * recvbuf, int nitems,
                         MPI_Datatype dtype, MPI_Op op, MPI_Comm comm) nogil:
     cdef int ierr
 
     if PRINT_DEBUG:
         with gil:
-            print_debug(f"MPI_Allreduce with {nitems} items", comm)
+            rid = random_id()
+            print_debug(f"MPI_Allreduce with {nitems} items", rid, comm)
+            start = perf_counter()
 
     # MPI Call
     ierr = libmpi.MPI_Allreduce(sendbuf, recvbuf, nitems, dtype, op, comm)
+
+    if PRINT_DEBUG:
+        with gil:
+            end = perf_counter()
[…]
