+++ .github/workflows/coin-flip.yml
@@ -1,13 +1,0 @@
-name: Flip a coin
-
-on:
-  workflow_dispatch:
-
-jobs:
-  flip:
-    runs-on: ubuntu-latest
-
-    steps:
-      - run: |
-          [[ $(($RANDOM % 2)) == 0 ]] && echo "heads" || echo "tails"
-        shell: bash
