### 2021-08-30 #3419 @dionhaefner
I'm done with incorporating all the suggestions. Thanks again for the review @Himscipy!
