+++ paper.bib
@@ -56,7 +56,7 @@ abstract = {The Message Passing Interface Forum (MPIF), with participation from 
   pages    = {1108-1115},
   year     = {2005},
   issn     = {0743-7315},
-  doi      = {https://doi.org/10.1016/j.jpdc.2005.03.010},
+  doi      = {10.1016/j.jpdc.2005.03.010},
   url      = {https://www.sciencedirect.com/science/article/pii/S0743731505000560},
   author   = {Lisandro Dalcín and Rodrigo Paz and Mario Storti},
   keywords = {Message passing, MPI, High level languages, Parallel Python},
