### 2021-08-31 #3419 @kyleniemeyer
Hi @dionhaefner, I'm the AEIC on duty this week, doing some final checks. It looks like the arXiv preprint by Degrave et al. that you cited has now been published as a peer-reviewed article, at https://www.frontiersin.org/articles/10.3389/fnbot.2019.00006/full.

Could you replace the citation to the arXiv preprint to that? The bibTex is:
```
@article{Degrave2019,
  doi = {10.3389/fnbot.2019.00006},
  year = {2019},
  month = mar,
  volume = {13},
  author = {Jonas Degrave and Michiel Hermans and Joni Dambre and Francis wyffels},
  title = {A Differentiable Physics Engine for Deep Learning in Robotics},
  journal = {Frontiers in Neurorobotics}
}
```
